+++ backend/enterprise/api/limit.go
@@ -33,7 +33,7 @@ const (
 )
 
 // PlanLimitValues is the plan limit value mapping.
-var PlanLimitValues = map[PlanLimit]map[api.PlanType]int64{
+var PlanLimitValues = map[PlanLimit]map[api.PlanType]int{
 	PlanLimitMaximumTask:        {},
 	PlanLimitMaximumEnvironment: {},
 	PlanLimitMaximumInstance:    {},
@@ -42,10 +42,10 @@ var PlanLimitValues = map[PlanLimit]map[api.PlanType]int64{
 
 type planLimitConfig struct {
 	Type                    api.PlanType `yaml:"type"`
-	MaximumInstanceCount    int64        `yaml:"maximumInstanceCount"`
-	MaximumSeatCount        int64        `yaml:"maximumSeatCount"`
-	MaximumTaskCount        int64        `yaml:"maximumTaskCount"`
-	MaximumEnvironmentCount int64        `yaml:"maximumEnvironmentCount"`
+	MaximumInstanceCount    int          `yaml:"maximumInstanceCount"`
+	MaximumSeatCount        int          `yaml:"maximumSeatCount"`
+	MaximumTaskCount        int          `yaml:"maximumTaskCount"`
+	MaximumEnvironmentCount int          `yaml:"maximumEnvironmentCount"`
 }
 
 type planConfg struct {
